Add server script and Dockerfile
Added a new server script, 'server.cjs.js' and Dockerfile to the application. The server script is being set up to facilitate local development by creating an Express server that serves the static files from the 'dist' folder. The Dockerfile has been introduced to containerize the application, making it easier to build, ship, and run the application in various environments. This includes both a build and production stage to ensure only necessary production dependencies are installed in the final image, reducing the overall size.