Switch build command to yarn and add express modules
The build command in Dockerfile is changed from npm to yarn to enhance build performance. Additionally, "express" and "serve-static" modules are added in package.json to aid in serving static files in production. The webpack configuration in vue.config.js intended for production has been commented out for the time being to avoid potential conflicts.