Add local build script and update end-point URL
Added a new `build-local.sh` script to facilitate local Docker image builds based on the Git branch or tag, considering the multi-arch nature of the Docker images. The script also cleans up the builder instance post-build. In addition to that, updated the URL endpoint in `GeneratorExport.vue` from "cloud-bulma-generate" to "localhost" for local development and testing.